requirements.txt
sqlalchemy
fastapi
uvicorn
fastapi_health
alembic
asyncpg
geoalchemy2
shapely
pillow
pydantic-settings
httpx


# Retry
tenacity

# Testing
pytest
pytest-cov
pytest_asyncio